\name{tli}
\alias{tli}
\title{Math scores from Texas Assessment of Academic Skills (TAAS)}
\description{
This data set contains math scores and demographic data of
100 randomly selected students participating in the Texas
Assessment of Academic Skills (TAAS).
}
\usage{data(tli)}
\format{A \code{data.frame} containing 100 observations with the
  following columns:
        \describe{
          \item{\code{grade}}{Year in school of student}
          \item{\code{sex}}{Gender of student}
          \item{\code{disadvg}}{Is the student economically disadvantaged?}
          \item{\code{ethnicty}}{Race of student}
          \item{\code{tlimth}}{Math score of student}
        }
}
\source{Texas Education Agency, \url{http://www.tea.state.tx.us}}
\keyword{datasets}
