\name{xtable-internal}
\alias{xtableLSMeans}
\alias{lagImpactMat}

\title{Internal xtable Functions}
\description{
  Internal functions for the package xtable
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
