\name{xtable-internal}
\alias{xtableMatharray}
\alias{xtableList}
\alias{print.xtableList}
\alias{xtableLSMeans}
\alias{sanitize}
\alias{sanitize.numbers}
\alias{sanitize.final}

\title{Internal xtable Functions}
\description{
  Internal functions for the package xtable
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }